+++ src/__tests__/pino.test.ts
@@ -33,6 +33,19 @@ describe('structured logger with pino', () => {
     log.info('this is a test message')
   })
 
+  it('should log a message with a prefix', () => {
+    expect.assertions(1)
+    const { log, mockedStream } = getLoggerInstance()
+
+    mockedStream.on('data', (data: Buffer) => {
+      const entry = JSON.parse(data.toString()) as unknown as Record<string, any>
+      expect(entry.msg).toBe('[testing] this is a test message')
+      mockedStream.destroy()
+    })
+
+    log.withPrefix('[testing]').info('this is a test message')
+  })
+
   it('should include context', () => {
     expect.assertions(2)
     const { log, mockedStream } = getLoggerInstance()